+++ public_tests_p1 4/README.txt
@@ -1,0 +1,1 @@
+For  n = 3, n = 4, input_1 is UNSOLVABLE while rest are SOLVABLE. For n = 5, all are SOLVABLE (come up with an UNSOLVABLE one and test - it must work also)
