+++ public_tests_p1 4/n_equals_4/input_2.txt
@@ -1,0 +1,4 @@
+12 15 6 10
+4 9 5 8
+14 13 0 2
+1 7 11 3
